Test query characteristics classification with filters
The beholder filter looks at the stream of queries going through the
filter and does simple classification of the characteristics of the
stream. The filter does locality-sensitive hashing with the help of a
hashtable.

The filter can be improved by altering the hashing function and increasing
the distribution of the generated hash values.